+++ discord_webhook/webhook.py
@@ -1,9 +1,12 @@
-import logging
-import json
-import time
 import datetime
+import json
+import logging
+import time
+from functools import partial
+from typing import Any, Dict, List, Optional, Tuple, Union, cast
+
 import requests
-from .webhook_exceptions import *
+from .webhook_exceptions import ColourNotInRangeException
 
 logger = logging.getLogger(__name__)
 
@@ -13,7 +16,33 @@ class DiscordWebhook:
     Webhook for Discord
     """
 
-    def __init__(self, url=None, content=None, username=None, avatar_url=None, **kwargs):
+    url: Optional[Union[str, List[str]]]
+    content: Optional[Union[str, bytes]]
+    username: Optional[str]
+    avatar_url: Optional[str]
+    tts: bool
+    files: Dict[str, Tuple[Optional[str], Union[bytes, str]]]
+    embeds: List[Dict[str, Any]]
+    proxies: Optional[Dict[str, str]]
+    allowed_mentions: List[str]
+    timeout: Optional[float]
+    rate_limit_retry: bool = False
+
+    def __init__(
+        self,
+        url: Optional[Union[str, List[str]]] = None,
+        *,
+        content: Optional[str] = None,
+        username: Optional[str] = None,
+        avatar_url: Optional[str] = None,
+        tts: bool = False,
+        files: Optional[Dict[str, Tuple[Optional[str], Union[bytes, str]]]] = None,
+        embeds: Optional[List[Dict[str, Any]]] = None,
+        proxies: Optional[Dict[str, str]] = None,
+        timeout: Optional[float] = None,
+        rate_limit_retry: bool = False,
+        allowed_mentions: Optional[List[str]] = None,
+    ) -> None:
         """
         Init Webhook for Discord
         ---------
@@ -22,7 +51,7 @@ class DiscordWebhook:
         :keyword ``username:`` override the default username of the webhook\n
         :keyword ``avatar_url:`` override the default avatar of the webhook\n
         :keyword ``tts:`` true if this is a TTS message\n
-        :keyword ``file``: to apply file(s) with message 
+        :keyword ``file``: to apply file(s) with message
         (For example: file=f.read() (here, f = variable that contain attachement path as "rb" mode))\n
         :keyword ``filename:`` apply custom file name on attached file content(s)\n
         :keyword ``embeds:`` list of embedded rich content\n
@@ -30,58 +59,64 @@ class DiscordWebhook:
         :keyword ``proxies:`` dict of proxies\n
         :keyword ``timeout:`` (optional) amount of seconds to wait for a response from Discord
         """
+        if embeds is None:
+            embeds = []
+        if files is None:
+            files = {}
+        if allowed_mentions is None:
+            allowed_mentions = []
         self.url = url
         self.content = content
         self.username = username
         self.avatar_url = avatar_url
-        self.tts = kwargs.get("tts", False)
-        self.files = kwargs.get("files", dict())
-        self.embeds = kwargs.get("embeds", [])
-        self.proxies = kwargs.get("proxies")
-        self.allowed_mentions = kwargs.get("allowed_mentions")
-        self.timeout = kwargs.get("timeout")
-        self.rate_limit_retry = kwargs.get("rate_limit_retry")
+        self.tts = tts
+        self.files = files
+        self.embeds = embeds
+        self.proxies = proxies
+        self.allowed_mentions = allowed_mentions
+        self.timeout = timeout
+        self.rate_limit_retry = rate_limit_retry
 
-    def add_file(self, file, filename):
+    def add_file(self, file: bytes, filename: str) -> None:
         """
         adds a file to the webhook
         :param file: file content
         :param filename: filename
         :return:
         """
-        self.files["_{}".format(filename)] = (filename, file)
+        self.files[f"_{filename}"] = (filename, file)
 
-    def add_embed(self, embed):
+    def add_embed(self, embed: Union[DiscordEmbed, Dict[str, Any]]) -> None:
         """
         adds an embedded rich content
         :param embed: embed object or dict
         """
         self.embeds.append(embed.__dict__ if isinstance(embed, DiscordEmbed) else embed)
 
-    def remove_embed(self, index):
+    def remove_embed(self, index: int) -> None:
         """
         removes embedded rich content from `self.embeds`
         :param index: index of embed in `self.embeds`
         """
         self.embeds.pop(index)
 
-    def remove_file(self, filename):
+    def remove_file(self, filename: str) -> None:
         """
         removes file from `self.files` using specified `filename` if it exists
         :param filename: filename
         """
-        filename = "_{}".format(filename)
+        filename = f"_{filename}"
         if filename in self.files:
             del self.files[filename]
 
-    def get_embeds(self):
+    def get_embeds(self) -> List[Dict[str, Any]]:
         """
         gets all self.embeds as list
         :return: self.embeds
         """
         return self.embeds
 
-    def set_proxies(self, proxies):
+    def set_proxies(self, proxies: Dict[str, str]) -> None:
         """
         sets proxies
         :param proxies: dict of proxies
@@ -89,7 +124,7 @@ class DiscordWebhook:
         """
         self.proxies = proxies
 
-    def set_content(self, content):
+    def set_content(self, content: str) -> None:
         """
         sets content
         :param content: content string
@@ -98,7 +133,7 @@ class DiscordWebhook:
         self.content = content
 
     @property
-    def json(self):
+    def json(self) -> Dict[str, Any]:
         """
         convert webhook data to json
         :return webhook data as json:
@@ -118,77 +153,71 @@ class DiscordWebhook:
             logger.error("webhook message is empty! set content or embed data")
         return data
 
-    def remove_embeds(self):
+    def remove_embeds(self) -> None:
         """
         Sets `self.embeds` to empty `list`.
         """
         self.embeds = []
 
-    def remove_files(self):
+    def remove_files(self) -> None:
         """
         Sets `self.files` to empty `dict`.
         """
         self.files = {}
 
-    def api_post_request(self, url):
+    def api_post_request(self, url: str) -> requests.Response:
         if bool(self.files) is False:
-            response = requests.post(url, json=self.json, proxies=self.proxies,
-                                     params={'wait': True},
-                                     timeout=self.timeout)
+            response = requests.post(
+                url,
+                json=self.json,
+                proxies=self.proxies,
+                params={"wait": True},
+                timeout=self.timeout,
+            )
         else:
             self.files["payload_json"] = (None, json.dumps(self.json))
-            response = requests.post(url, files=self.files,
-                                     proxies=self.proxies,
-                                     timeout=self.timeout)
-
+            response = requests.post(
+                url,
+                files=self.files,
+                proxies=self.proxies,
+                timeout=self.timeout,
+            )
         return response
 
-    def execute(self, remove_embeds=False, remove_files=False):
+    def execute(
+        self,
+        remove_embeds: bool = False,
+        remove_files: bool = False,
+    ) -> Union[List[requests.Response], requests.Response]:
         """
         executes the Webhook
         :param remove_embeds: if set to True, calls `self.remove_embeds()` to empty `self.embeds` after webhook is executed
         :param remove_files: if set to True, calls `self.remove_files()` to empty `self.files` after webhook is executed
         :return: Webhook response
         """
-        webhook_urls = self.url if isinstance(self.url, list) else [self.url]
+        if isinstance(self.url, str):
+            webhook_urls = [self.url]
         urls_len = len(webhook_urls)
         responses = []
         for i, url in enumerate(webhook_urls):
             response = self.api_post_request(url)
             if response.status_code in [200, 204]:
-                logger.debug(
-                    "[{index}/{length}] Webhook executed".format(
-                        index=i + 1, length=urls_len
-                    )
-                )
+                logger.debug(f"[{i+1}/{urls_len}] Webhook executed")
             elif response.status_code == 429 and self.rate_limit_retry:
                 while response.status_code == 429:
-                    errors = json.loads(
-                        response.content.decode('utf-8'))
-                    wh_sleep = (int(errors['retry_after']) / 1000) + 0.15
+                    errors = json.loads(response.content.decode("utf-8"))
+                    wh_sleep = (int(errors["retry_after"]) / 1000) + 0.15
                     time.sleep(wh_sleep)
                     logger.error(
-                        "Webhook rate limited: sleeping for {wh_sleep} "
-                        "seconds...".format(
-                            wh_sleep=wh_sleep
-                        )
+                        f"Webhook rate limited: sleeping for {wh_sleep} " "seconds..."
                     )
                     response = self.api_post_request(url)
                     if response.status_code in [200, 204]:
-                        logger.debug(
-                            "[{index}/{length}] Webhook executed".format(
-                                index=i + 1, length=urls_len
-                            )
-                        )
+                        logger.debug(f"[{i+1}/{urls_len}] Webhook executed")
                         break
             else:
                 logger.error(
-                    "[{index}/{length}] Webhook status code {status_code}: {content}".format(
-                        index=i + 1,
-                        length=urls_len,
-                        status_code=response.status_code,
-                        content=response.content.decode("utf-8"),
-                    )
+                    f"[{i+1}/{urls_len}] Webhook status code {response.status_code}: {response.content.decode('utf-8')}"
                 )
             responses.append(response)
         if remove_embeds:
@@ -197,92 +226,92 @@ class DiscordWebhook:
             self.remove_files()
         return responses[0] if len(responses) == 1 else responses
 
-    def edit(self, sent_webhook):
+    def edit(
+        self,
+        sent_webhook: Union[List[requests.Response], requests.Response],
+    ) -> Union[List[requests.Response], requests.Response]:
         """
         edits the webhook passed as a response
         :param sent_webhook: webhook.execute() response
         :return: Another webhook response
         """
-        sent_webhook = sent_webhook if isinstance(sent_webhook, list) else [sent_webhook]
-        webhook_len = len(sent_webhook)
-        responses = []
+        if not isinstance(sent_webhook, list):
+            sent_webhook = cast(List[requests.Response], [sent_webhook])
+        responses: List[requests.Response] = []
         for i, webhook in enumerate(sent_webhook):
-            previous_sent_message_id = json.loads(webhook.content.decode('utf-8'))['id']
-            url = webhook.url.split('?')[0] + '/messages/' + str(previous_sent_message_id)  # removes any query params
+            assert isinstance(webhook.content, bytes)
+            assert isinstance(webhook.url, str)
+            previous_sent_message_id = json.loads(webhook.content.decode("utf-8"))["id"]
+            url = (
+                webhook.url.split("?")[0] + "/messages/" + str(previous_sent_message_id)
+            )
+            # removes any query params
             if bool(self.files) is False:
-                patch_kwargs = {'json': self.json, 'proxies': self.proxies, 'params': {'wait': True}, 'timeout': self.timeout}
+                request = partial(
+                    requests.patch,
+                    url,
+                    json=self.json,
+                    proxies=self.proxies,
+                    params={"wait": True},
+                    timeout=self.timeout,
+                )
             else:
                 self.files["payload_json"] = (None, json.dumps(self.json))
-                patch_kwargs = {'files': self.files, 'proxies': self.proxies, 'timeout': self.timeout}
-            response = requests.patch(url, **patch_kwargs)
-            if response.status_code in [200, 204]:
-                logger.debug(
-                    "[{index}/{length}] Webhook edited".format(
-                        index=i + 1,
-                        length=webhook_len,
-                    )
+                request = partial(
+                    requests.patch,
+                    url,
+                    files=self.files,
+                    proxies=self.proxies,
+                    timeout=self.timeout,
                 )
+            response = request()
+            if response.status_code in [200, 204]:
+                logger.debug(f"[{i+1}/{len(sent_webhook)}] Webhook edited")
             elif response.status_code == 429 and self.rate_limit_retry:
                 while response.status_code == 429:
-                    errors = json.loads(response.content.decode('utf-8'))
-                    wh_sleep = (int(errors['retry_after']) / 1000) + 0.15
+                    errors = json.loads(response.content.decode("utf-8"))
+                    wh_sleep = (int(errors["retry_after"]) / 1000) + 0.15
                     time.sleep(wh_sleep)
                     logger.error(
-                        "Webhook rate limited: sleeping for {wh_sleep} "
-                        "seconds...".format(
-                            wh_sleep=wh_sleep
-                        )
+                        f"Webhook rate limited: sleeping for {wh_sleep} seconds..."
                     )
-                    response = requests.patch(url, **patch_kwargs)
+                    response = request()
                     if response.status_code in [200, 204]:
-                        logger.debug(
-                            "[{index}/{length}] Webhook edited".format(
-                                index=i + 1,
-                                length=webhook_len,
-                            )
-                        )
+                        logger.debug(f"[{i + 1}/{len(sent_webhook)}] Webhook edited")
                         break
             else:
                 logger.error(
-                    "[{index}/{length}] Webhook status code {status_code}: {content}".format(
-                        index=i + 1,
-                        length=webhook_len,
-                        status_code=response.status_code,
-                        content=response.content.decode("utf-8"),
-                    )
+                    f"[{i+1}/{len(sent_webhook)}] Webhook status code {response.status_code}: {response.content.decode('utf-8')}"
                 )
             responses.append(response)
         return responses[0] if len(responses) == 1 else responses
 
-    def delete(self, sent_webhook):
+    def delete(
+        self, sent_webhook: Union[List["DiscordWebhook"], "DiscordWebhook"]
+    ) -> Union[List[requests.Response], requests.Response]:
         """
         deletes the webhook passed as a response
         :param sent_webhook: webhook.execute() response
         :return: Response
         """
-        sent_webhook = sent_webhook if isinstance(sent_webhook, list) else [sent_webhook]
-        webhook_len = len(sent_webhook)
-        responses = []
+        if not isinstance(sent_webhook, list):
+            sent_webhook = cast(List[DiscordWebhook], [sent_webhook])
+        responses: List[requests.Response] = []
         for i, webhook in enumerate(sent_webhook):
-            url = webhook.url.split('?')[0]  # removes any query params
-            previous_sent_message_id = json.loads(webhook.content.decode('utf-8'))['id']
-            response = requests.delete(url + '/messages/' + str(previous_sent_message_id), proxies=self.proxies,
-                                       timeout=self.timeout)
+            assert isinstance(webhook.content, bytes)
+            assert isinstance(webhook.content, str)
+            url = webhook.url.split("?")[0]  # removes any query params
+            previous_sent_message_id = json.loads(webhook.content.decode("utf-8"))["id"]
+            response = requests.delete(
+                url + "/messages/" + str(previous_sent_message_id),
+                proxies=self.proxies,
+                timeout=self.timeout,
+            )
             if response.status_code in [200, 204]:
-                logger.debug(
-                    "[{index}/{length}] Webhook deleted".format(
-                        index=i + 1,
-                        length=webhook_len,
-                    )
-                )
+                logger.debug(f"[{i+1}/{len(sent_webhook)}] Webhook deleted")
             else:
                 logger.error(
-                    "[{index}/{length}] Webhook status code {status_code}: {content}".format(
-                        index=i + 1,
-                        length=webhook_len,
-                        status_code=response.status_code,
-                        content=response.content.decode("utf-8"),
-                    )
+                    f"[{i+1}/{len(sent_webhook)}] Webhook status code {response.status_code}: {response.content.decode('utf-8')}"
                 )
             responses.append(response)
         return responses[0] if len(responses) == 1 else responses
@@ -293,7 +322,27 @@ class DiscordEmbed:
     Discord Embed
     """
 
-    def __init__(self, title=None, description=None, hex_color='33ccff', **kwargs):
+    title: Optional[str]
+    description: Optional[str]
+    url: Optional[str]
+    timestamp: Optional[str]
+    color: Optional[int]
+    hex_color: Optional[str]
+    footer: Optional[Dict[str, Optional[str]]]
+    image: Optional[Dict[str, Optional[Union[str, int]]]]
+    thumbnail: Optional[Union[str, Dict[str, Optional[Union[str, int]]]]]
+    video: Optional[Union[str, Dict[str, Optional[Union[str, int]]]]]
+    provider: Optional[Dict[str, Any]]
+    author: Optional[Dict[str, Optional[str]]]
+    fields: List[Dict[str, Optional[Any]]]
+
+    def __init__(
+        self,
+        title: Optional[str] = None,
+        description: Optional[str] = None,
+        hex_color: str = "33ccff",
+        **kwargs: Any,
+    ) -> None:
         """
         Init Discord Embed
         -----------
@@ -313,11 +362,8 @@ class DiscordEmbed:
         """
         self.title = title
         self.description = description
-        self.url = kwargs.get("url")
-        self.timestamp = kwargs.get("timestamp")
-        self.color = kwargs.get("color")
-        if self.color:
-            self.set_color(self.color)
+        self.url = cast(str, kwargs.get("url"))
+        self.timestamp = cast(str, kwargs.get("timestamp"))
         self.hex_color = hex_color
         self.footer = kwargs.get("footer")
         self.image = kwargs.get("image")
@@ -326,29 +372,30 @@ class DiscordEmbed:
         self.provider = kwargs.get("provider")
         self.author = kwargs.get("author")
         self.fields = kwargs.get("fields", [])
+        self.color = cast(Optional[int], kwargs.get("color"))
 
-    def set_title(self, title):
+    def set_title(self, title: str) -> None:
         """
         set title of embed
         :param title: title of embed
         """
         self.title = title
 
-    def set_description(self, description):
+    def set_description(self, description: str) -> None:
         """
         set description of embed
         :param description: description of embed
         """
         self.description = description
 
-    def set_url(self, url):
+    def set_url(self, url: str) -> None:
         """
         set url of embed
         :param url: url of embed
         """
         self.url = url
 
-    def set_timestamp(self, timestamp=None):
+    def set_timestamp(self, timestamp: Optional[float] = None) -> None:
         """
         set timestamp of embed content
         :param timestamp: (optional) timestamp of embed content
@@ -357,7 +404,7 @@ class DiscordEmbed:
             timestamp = time.time()
         self.timestamp = str(datetime.datetime.utcfromtimestamp(timestamp))
 
-    def set_color(self, color):
+    def set_color(self, color: Union[str, int]) -> None:
         """
         set color code of the embed as decimal(int) or hex(string)
         :param color: color code of the embed as decimal(int) or hex(string)
@@ -366,7 +413,7 @@ class DiscordEmbed:
         if self.color not in range(16777216):
             raise ColourNotInRangeException(color)
 
-    def set_footer(self, **kwargs):
+    def set_footer(self, **kwargs: str) -> None:
         """
         set footer information of embed
         :keyword text: footer text
@@ -379,7 +426,7 @@ class DiscordEmbed:
             "proxy_icon_url": kwargs.get("proxy_icon_url"),
         }
 
-    def set_image(self, **kwargs):
+    def set_image(self, **kwargs: Union[str, int]) -> None:
         """
         set image of embed
         :keyword url: source url of image (only supports http(s) and attachments)
@@ -388,13 +435,13 @@ class DiscordEmbed:
         :keyword width: width of image
         """
         self.image = {
-            "url": kwargs.get("url"),
-            "proxy_url": kwargs.get("proxy_url"),
-            "height": kwargs.get("height"),
-            "width": kwargs.get("width"),
+            "url": cast(Optional[str], kwargs.get("url")),
+            "proxy_url": cast(Optional[str], kwargs.get("proxy_url")),
+            "height": cast(Optional[int], kwargs.get("height")),
+            "width": cast(Optional[int], kwargs.get("width")),
         }
 
-    def set_thumbnail(self, **kwargs):
+    def set_thumbnail(self, **kwargs: Union[str, int]) -> None:
         """
         set thumbnail of embed
         :keyword url: source url of thumbnail (only supports http(s) and attachments)
@@ -403,13 +450,13 @@ class DiscordEmbed:
         :keyword width: width of thumbnail
         """
         self.thumbnail = {
-            "url": kwargs.get("url"),
-            "proxy_url": kwargs.get("proxy_url"),
-            "height": kwargs.get("height"),
-            "width": kwargs.get("width"),
+            "url": cast(Optional[str], kwargs.get("url")),
+            "proxy_url": cast(Optional[str], kwargs.get("proxy_url")),
+            "height": cast(Optional[int], kwargs.get("height")),
+            "width": cast(Optional[str], kwargs.get("width")),
         }
 
-    def set_video(self, **kwargs):
+    def set_video(self, **kwargs: Union[str, int]) -> None:
         """
         set video of embed
         :keyword url: source url of video
@@ -417,12 +464,12 @@ class DiscordEmbed:
         :keyword width: width of video
         """
         self.video = {
-            "url": kwargs.get("url"),
-            "height": kwargs.get("height"),
-            "width": kwargs.get("width"),
+            "url": cast(Optional[str], kwargs.get("url")),
+            "height": cast(Optional[int], kwargs.get("height")),
+            "width": cast(Optional[int], kwargs.get("width")),
         }
 
-    def set_provider(self, **kwargs):
+    def set_provider(self, **kwargs: str) -> None:
         """
         set provider of embed
         :keyword name: name of provider
@@ -433,7 +480,7 @@ class DiscordEmbed:
             "url": kwargs.get("url"),
         }
 
-    def set_author(self, **kwargs):
+    def set_author(self, **kwargs: str) -> None:
         """
         set author of embed
         :keyword name: name of author
@@ -448,7 +495,7 @@ class DiscordEmbed:
             "proxy_icon_url": kwargs.get("proxy_icon_url"),
         }
 
-    def add_embed_field(self, **kwargs):
+    def add_embed_field(self, **kwargs: Union[str, bool]) -> None:
         """
         set field of embed
         :keyword name: name of the field
@@ -463,14 +510,14 @@ class DiscordEmbed:
             }
         )
 
-    def del_embed_field(self, index):
+    def del_embed_field(self, index: int) -> None:
         """
         remove field from `self.fields`
         :param index: index of field in `self.fields`
         """
         self.fields.pop(index)
 
-    def get_embed_fields(self):
+    def get_embed_fields(self) -> List[Dict[str, Optional[Any]]]:
         """
         get all `self.fields` as list
         :return: `self.fields`