+++ tests/test_discordwebhook.py
@@ -1,4 +1,5 @@
 import unittest
+
 from discord_webhook.webhook import DiscordWebhook